Deleagate Search Menu Responsibility
